[DZ-134] Refator : Refactor inside of 'common' package
