+++ src/main/java/com/dungzi/backend/domain/room/api/RoomController.java
@@ -5,8 +5,8 @@ import com.dungzi.backend.domain.room.dto.*;
 import com.dungzi.backend.domain.room.dto.RoomResponseDto;
 import com.dungzi.backend.domain.user.application.AuthService;
 import com.dungzi.backend.domain.user.domain.User;
-import com.dungzi.backend.global.common.CommonCode;
-import com.dungzi.backend.global.common.CommonResponse;
+import com.dungzi.backend.global.common.response.code.CommonCode;
+import com.dungzi.backend.global.common.response.CommonResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.web.bind.annotation.*;
