+++ src/main/java/com/dungzi/backend/domain/univ/api/UnivController.java
@@ -3,8 +3,8 @@ package com.dungzi.backend.domain.univ.api;
 import com.dungzi.backend.domain.univ.application.UnivService;
 import com.dungzi.backend.domain.univ.dto.UnivRequestDto;
 import com.dungzi.backend.domain.univ.dto.UnivResponseDto;
-import com.dungzi.backend.global.common.CommonCode;
-import com.dungzi.backend.global.common.CommonResponse;
+import com.dungzi.backend.global.common.response.code.CommonCode;
+import com.dungzi.backend.global.common.response.CommonResponse;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
