+++ src/main/java/com/dungzi/backend/domain/user/api/AuthController.java
@@ -9,12 +9,12 @@ import com.dungzi.backend.domain.user.application.KakaoService;
 import com.dungzi.backend.domain.user.domain.User;
 import com.dungzi.backend.domain.user.dto.UserRequestDto;
 import com.dungzi.backend.domain.user.dto.UserAuthResponseDto;
-import com.dungzi.backend.global.common.CommonCode;
-import com.dungzi.backend.global.common.CommonResponse;
-import com.dungzi.backend.global.common.error.AuthException;
-import com.dungzi.backend.global.common.error.AuthErrorCode;
-import com.dungzi.backend.global.common.error.ValidErrorCode;
-import com.dungzi.backend.global.common.error.ValidException;
+import com.dungzi.backend.global.common.response.code.CommonCode;
+import com.dungzi.backend.global.common.response.CommonResponse;
+import com.dungzi.backend.global.common.response.exception.AuthException;
+import com.dungzi.backend.global.common.response.code.AuthErrorCode;
+import com.dungzi.backend.global.common.response.code.ValidErrorCode;
+import com.dungzi.backend.global.common.response.exception.ValidException;
 import io.swagger.v3.oas.annotations.Operation;
 import io.swagger.v3.oas.annotations.responses.ApiResponse;
 import io.swagger.v3.oas.annotations.responses.ApiResponses;
