+++ src/main/java/com/dungzi/backend/domain/user/application/EmailService.java
@@ -1,6 +1,6 @@
 package com.dungzi.backend.domain.user.application;
 
-import com.dungzi.backend.global.common.error.AuthException;
+import com.dungzi.backend.global.common.response.exception.AuthException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
@@ -10,8 +10,6 @@ import org.springframework.stereotype.Service;
 import javax.mail.MessagingException;
 import javax.mail.internet.InternetAddress;
 import javax.mail.internet.MimeMessage;
-import java.io.BufferedWriter;
-import java.io.OutputStreamWriter;
 import java.io.UnsupportedEncodingException;
 import java.util.Random;
