+++ src/main/java/com/dungzi/backend/global/common/response/exception/DefaultException.java
@@ -1,6 +1,6 @@
-package com.dungzi.backend.global.common.error;
+package com.dungzi.backend.global.common.response.exception;
 
-import com.dungzi.backend.global.common.Code;
+import com.dungzi.backend.global.common.response.code.Code;
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
